pyproject.toml
[tool.poetry]
name = "freedom-gate"
version = "0.1.0"
description = ""
authors = ["None"]
license = "Free"
readme = "README.md"
packages = [{include = "freedom_gate"}]

[tool.poetry.dependencies]
python = "^3.9"
pythonping = "^1.1.3"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
html5lib = "^1.1"
pymongo = "^4.2.0"
loguru = "^0.6.0"
PyYAML = "^6.0"
playwright = "^1.26.0"
fake-useragent = "^0.1.11"
httpx = "^0.23.0"
click = "^8.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
